Add a singleton location controller
